+++ src/twebapplication.h
@@ -64,8 +64,13 @@ public:
     QString systemLogFilePath() const;
     QString accessLogFilePath() const;
     QString sqlQueryLogFilePath() const;
+#if QT_VERSION < 0x060000
     QTextCodec *codecForInternal() const { return _codecInternal; }
     QTextCodec *codecForHttpOutput() const { return _codecHttp; }
+#else
+    QStringConverter::Encoding encodingForInternal() const { return _encodingInternal; }
+    QStringConverter::Encoding encodingForHttpOutput() const { return _encodingHttp; }
+#endif
     int applicationServerId() const { return _appServerId; }
     QThread *databaseContextMainThread() const;
     const QVariantMap &getConfig(const QString &configName);
@@ -98,8 +103,13 @@ private:
     QVariantMap _loggerSetting;
     QVariantMap _validationSetting;
     QVariantMap _mediaTypes;
+#if QT_VERSION < 0x060000
     QTextCodec *_codecInternal {nullptr};
     QTextCodec *_codecHttp {nullptr};
+#else
+    QStringConverter::Encoding _encodingInternal {QStringConverter::Utf8};
+    QStringConverter::Encoding _encodingHttp {QStringConverter::Utf8};
+#endif
     int _appServerId {-1};
     QBasicTimer _timer;
     mutable MultiProcessingModule _mpm {Invalid};