+++ tools/tspawn/controllergenerator.cpp
@@ -62,7 +62,6 @@ constexpr auto CONTROLLER_SOURCE_FILE_TEMPLATE = "#include \"%name%controller.h\
                                                  "            render();\n"
                                                  "        }\n"
                                                  "        break;\n"
-                                                 "\n"
                                                  "    default:\n"
                                                  "        renderErrorResponse(Tf::NotFound);\n"
                                                  "        break;\n"
@@ -244,7 +243,11 @@ bool ControllerGenerator::generate(const QString &dstDir) const
             {"var1", convMethod()->value(pair.second).arg(fieldNameToVariableName(pair.first))},
             {"code1", sessInsertStr},
             {"code2", sessGetStr},
+#if QT_VERSION < 0x060000
             {"type", QString::fromLatin1(QMetaType::typeName(pair.second))},
+#else
+            {"type", QString::fromLatin1(QMetaType(pair.second).name())},
+#endif
             {"rev", revStr},
             {"id", fieldNameToVariableName(pair.first)},
             {"condition", conditionString()->value(pair.second).arg(fieldNameToVariableName(pair.first))},