+++ tools/tspawn/main.cpp
@@ -15,6 +15,8 @@
 #include "mongocommand.h"
 #include "mongoobjgenerator.h"
 #include "otamagenerator.h"
+#include "vueservicegenerator.h"
+#include "vueerbgenerator.h"
 #include "projectfilegenerator.h"
 #include "sqlobjgenerator.h"
 #include "tableschema.h"
@@ -24,7 +26,9 @@
 #include "apicontrollergenerator.h"
 #include "apiservicegenerator.h"
 #include <QtCore>
-#include <QTextCodec>
+#if QT_VERSION < 0x060000
+# include <QTextCodec>
+#endif
 #include <random>
 #ifndef Q_CC_MSVC
 #include <unistd.h>
@@ -203,7 +207,6 @@ static QSettings appSettings(appIni, QSettings::IniFormat);
 static QSettings devSettings(devIni, QSettings::IniFormat);
 static QString templateSystem;
 
-
 static void usage()
 {
     std::printf("usage: tspawn <subcommand> [args]\n\n"
@@ -253,7 +256,7 @@ static QStringList rmfiles(const QStringList &files, bool &allRemove, bool &quit
         for (;;) {
             std::printf("  remove  %s? [ynaqh] ", qUtf8Printable(QDir::cleanPath(file.fileName())));
 
-            QString line = stream.readLine();
+            QString line = stream.readLine().trimmed();
             if (line.isNull())
                 break;
 
@@ -515,6 +518,31 @@ static void printSuccessMessage(const QString &model)
 }
 
 
+static bool isVueEnabled()
+{
+    static int vueEnable = -1;
+
+    if (vueEnable < 0) {
+        std::printf("\n");
+        QTextStream stream(stdin);
+        for (;;) {
+            std::printf(" Create sources for vue.js? [y/n] ");
+            QString line = stream.readLine().trimmed();
+
+            const QChar c = line[0];
+            if (c == 'Y' || c == 'y') {
+                vueEnable = 1;
+                break;
+            } else if (c == 'N' || c == 'n') {
+                vueEnable = 0;
+                break;
+            }
+        }
+    }
+    return (bool)vueEnable;
+}
+
+
 int main(int argc, char *argv[])
 {
     QCoreApplication app(argc, argv);
@@ -610,10 +638,12 @@ int main(int argc, char *argv[])
             return 2;
         }
 
+#if QT_VERSION < 0x060000
         // Sets codec
         QTextCodec *codec = QTextCodec::codecForName(appSettings.value("InternalEncoding").toByteArray().trimmed());
         codec = (codec) ? codec : QTextCodec::codecForLocale();
         QTextCodec::setCodecForLocale(codec);
+#endif
 
         // ERB or Otama
         templateSystem = devSettings.value("TemplateSystem").toString().toLower();
@@ -644,8 +674,13 @@ int main(int argc, char *argv[])
                 return 2;
             }
 
-            ServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
-            svrgen.generate(D_MODELS);
+            if (isVueEnabled()) {
+                VueServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
+                svrgen.generate(D_MODELS);
+            } else {
+                ServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
+                svrgen.generate(D_MODELS);
+            }
             break;
         }
 
@@ -665,8 +700,13 @@ int main(int argc, char *argv[])
                 return 2;
             }
 
-            ServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
-            svrgen.generate(D_MODELS);
+            if (isVueEnabled()) {
+                VueServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
+                svrgen.generate(D_MODELS);
+            } else {
+                ServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
+                svrgen.generate(D_MODELS);
+            }
             break;
         }
 
@@ -690,8 +730,13 @@ int main(int argc, char *argv[])
                 return 2;
             }
 
-            ServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
-            success &= svrgen.generate(D_MODELS);
+            if (isVueEnabled()) {
+                VueServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
+                success &= svrgen.generate(D_MODELS);
+            } else {
+                ServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
+                success &= svrgen.generate(D_MODELS);
+            }
 
             ControllerGenerator crtlgen(modelgen.model(), modelgen.fieldList(), modelgen.primaryKeyIndex(), modelgen.lockRevisionIndex());
             success &= crtlgen.generate(D_CTRLS);
@@ -724,8 +769,13 @@ int main(int argc, char *argv[])
                 return 2;
             }
 
-            ServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
-            svrgen.generate(D_MODELS);
+            if (isVueEnabled()) {
+                VueServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
+                svrgen.generate(D_MODELS);
+            } else {
+                ServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
+                svrgen.generate(D_MODELS);
+            }
             break;
         }
 
@@ -793,22 +843,36 @@ int main(int argc, char *argv[])
                 return 2;
             }
 
-            ServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
-            svrgen.generate(D_MODELS);
-
             ControllerGenerator crtlgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
             success &= crtlgen.generate(D_CTRLS);
 
-            // Generates view files of the specified template system
-            if (templateSystem == "otama") {
-                OtamaGenerator viewgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.autoValueIndex());
-                viewgen.generate(D_VIEWS);
-            } else if (templateSystem == "erb") {
-                ErbGenerator viewgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.autoValueIndex());
-                viewgen.generate(D_VIEWS);
+            if (isVueEnabled()) {
+                VueServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
+                svrgen.generate(D_MODELS);
+
+                // Generates view files of the specified template system
+                if (templateSystem == "erb") {
+                    VueErbGenerator viewgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.autoValueIndex());
+                    viewgen.generate(D_VIEWS);
+                } else {
+                    qCritical("Invalid template system specified: %s", qUtf8Printable(templateSystem));
+                    return 2;
+                }
             } else {
-                qCritical("Invalid template system specified: %s", qUtf8Printable(templateSystem));
-                return 2;
+                ServiceGenerator svrgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.lockRevisionIndex());
+                svrgen.generate(D_MODELS);
+
+                // Generates view files of the specified template system
+                if (templateSystem == "otama") {
+                    OtamaGenerator viewgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.autoValueIndex());
+                    viewgen.generate(D_VIEWS);
+                } else if (templateSystem == "erb") {
+                    ErbGenerator viewgen(modelgen.model(), modelgen.fieldList(), pkidx, modelgen.autoValueIndex());
+                    viewgen.generate(D_VIEWS);
+                } else {
+                    qCritical("Invalid template system specified: %s", qUtf8Printable(templateSystem));
+                    return 2;
+                }
             }
 
             if (success) {